Tool name and command standardization
We've been using the tool name and the actual command in different
ways among the different check scripts.  This standardizes them to use
the same style, and the same behavior.

While it would make sense to call the tools by their names because
that would be applicable to any tool (be it Python based on or) there
are situations on too many platforms where, given the installation of
the tool using PIP, the scripts are not readily available on the PATH.

Even worse, is that different versions may be picked up from outside a
virtual environment, like from system packages.  This change
standardizes on using the Python module entrypoint for the tools
written Python, which will get the right tool if inside a virtual
environment becase "python3" will be already set to the active one.
A similar change was done for "selftests/style.sh" on Avocado itself.

This standardizes the following:

 * TOOL_NAME is the descriptive name that is given to humans
 * TOOL_CMD is the actual command that will be called

Reference: https://github.com/avocado-framework/avocado/commit/77ca585f776dba3d07df81eb41d51828249d54a9
Signed-off-by: Cleber Rosa <crosa@redhat.com>